requirements.txt
Fabric==1.8.1
django==1.6
South==0.8.4
argparse==1.2.1
django-ajax
django-crispy-forms==1.4.0
django-jsonfield==0.9.12
django-localflavor==1.0
django-registration==1.0
ecdsa==0.10
eventlog==0.8.0
ipdb==0.8
ipython==1.1.0
paramiko==1.12.1
psycopg2==2.5.2
pycrypto==2.6.1
wsgiref==0.1.2
